feat: Introduce vault management with password count tracking and messaging integration

+++ self-hosted/mossy-vault/src/main/kotlin/pl/dayfit/mossyvault/messaging/consumer/DeletePasswordHandler.kt
@@ -3,13 +3,20 @@ package pl.dayfit.mossyvault.messaging.consumer
 import org.springframework.messaging.simp.stomp.StompFrameHandler
 import org.springframework.messaging.simp.stomp.StompHeaders
 import org.springframework.stereotype.Component
+import pl.dayfit.mossyvault.configuration.StompEndpoints
 import pl.dayfit.mossyvault.dto.request.DeletePasswordRequestDto
+import pl.dayfit.mossyvault.dto.response.DeletePasswordVaultResponseDto
+import pl.dayfit.mossyvault.repository.PasswordEntryRepository
 import pl.dayfit.mossyvault.service.PasswordEntryService
+import pl.dayfit.mossyvault.service.StompSessionRegistry
 import java.lang.reflect.Type
+import kotlin.jvm.optionals.getOrNull
 
 @Component
 class DeletePasswordHandler(
-    private val passwordEntryService: PasswordEntryService
+    private val passwordEntryService: PasswordEntryService,
+    private val stompSessionRegistry: StompSessionRegistry,
+    private val passwordEntryRepository: PasswordEntryRepository
 ) : StompFrameHandler {
     private val logger = org.slf4j.LoggerFactory.getLogger(DeletePasswordHandler::class.java)
 
@@ -25,6 +32,22 @@ class DeletePasswordHandler(
             return
         }
 
+        val passwordEntry = passwordEntryRepository.findById(requestDto.passwordId)
+            .getOrNull()
+
+        if (passwordEntry == null) {
+            logger.warn("Password entry not found, id={}, ignoring", requestDto.passwordId)
+            return
+        }
+
         passwordEntryService.delete(requestDto)
+
+        stompSessionRegistry.send(
+            StompEndpoints.USER_PASSWORD_DELETED,
+            DeletePasswordVaultResponseDto(
+                passwordEntry.domain,
+                passwordEntry.id!!
+            )
+        )
     }
 }
